feat: 로그인 기능 구현 후 반영

+++ src/main/java/com/hanamja/moa/api/service/point_history/PointHistoryService.java
@@ -5,8 +5,10 @@ import com.hanamja.moa.api.dto.util.DataResponseDto;
 import com.hanamja.moa.api.entity.point_history.PointHistory;
 import com.hanamja.moa.api.entity.point_history.PointHistoryRepository;
 import com.hanamja.moa.api.entity.user.User;
+import com.hanamja.moa.api.entity.user.UserAccount.UserAccount;
 import com.hanamja.moa.api.entity.user.UserRepository;
 import com.hanamja.moa.exception.custom.NotFoundException;
+import com.hanamja.moa.exception.custom.UnauthorizedException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -22,9 +24,8 @@ public class PointHistoryService {
     private final UserRepository userRepository;
     private final PointHistoryRepository pointHistoryRepository;
 
-    public DataResponseDto<List<PointHistoryInfoResponseDto>> getHistoryList() {
-        // TODO: 로그인 구현 후 @AuthenticationPrincipal User user 추가 필요
-        User user = userRepository.findById(1L).orElseThrow();
+    public DataResponseDto<List<PointHistoryInfoResponseDto>> getHistoryList(UserAccount userAccount) {
+        User user = validateUser(userAccount.getUserId());
 
         return DataResponseDto.<List<PointHistoryInfoResponseDto>>builder()
                 .data(pointHistoryRepository
@@ -51,9 +52,8 @@ public class PointHistoryService {
 //        );
 //    }
 
-    public PointHistoryInfoResponseDto removeHistory(Long historyId) {
-        // TODO: 로그인 구현 후 @AuthenticationPrincipal User user 추가 필요
-        User user = userRepository.findById(1L).orElseThrow();
+    public PointHistoryInfoResponseDto removeHistory(UserAccount userAccount, Long historyId) {
+        User user = validateUser(userAccount.getUserId());
 
         PointHistory pointHistory = pointHistoryRepository.findById(historyId).orElseThrow(
                 () -> NotFoundException
@@ -63,8 +63,26 @@ public class PointHistoryService {
                         .build()
         );
 
+        if (!pointHistory.getOwner().getId().equals(user.getId())) {
+            throw UnauthorizedException
+                    .builder()
+                    .httpStatus(HttpStatus.FORBIDDEN)
+                    .message("본인의 포인트 적립 내역만 삭제할 수 있습니다.")
+                    .build();
+        }
+
         pointHistoryRepository.deleteById(pointHistory.getId());
 
         return PointHistoryInfoResponseDto.from(pointHistory);
+    }
+
+    private User validateUser(Long userId) {
+        return userRepository.findById(userId).orElseThrow(
+                () -> NotFoundException
+                        .builder()
+                        .httpStatus(HttpStatus.BAD_REQUEST)
+                        .message("유효하지 않은 사용자입니다.")
+                        .build()
+        );
     }
 }
